backend/requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Zoom SDK/API
zoomus==1.1.7
pyzoom==0.1.1

# HeyGen API (using requests until official SDK)
requests==2.31.0
httpx==0.26.0

# Deepgram SDK
deepgram-sdk==3.2.0

# Vector database for RAG
chromadb==0.4.22
sentence-transformers==2.3.1

# Document processing
PyPDF2==3.0.1
python-pptx==0.6.23
python-docx==1.1.0
markdown==3.5.2

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
